+++ headers/Ray.hpp
@@ -17,7 +17,7 @@ public:
 
 	Vec3 origin() const {return origin_vec;}
 	Vec3 direction() const {return direction_vec;}
-	Vec3 point_at_parameter(double t) const {return origin_vec + t*direction_vec; }
+	Vec3 point_at_parameter(float t) const {return origin_vec + t*direction_vec; }
 };
 
 #endif
